* Don't use `Exception` for raising exceptions as it cannot be caught specifically without catching all exceptions * Use a base class for Sharepoint errors so they can be caught more specifically
